ref: accelerator names
