+++ pytorch_lightning/accelerators/__init__.py
@@ -1,14 +1,14 @@
-from pytorch_lightning.accelerators.cpu_backend import CPUBackend
-from pytorch_lightning.accelerators.ddp2_backend import DDP2Backend
-from pytorch_lightning.accelerators.ddp_backend import DDPBackend
-from pytorch_lightning.accelerators.ddp_spawn_backend import DDPSpawnBackend
-from pytorch_lightning.accelerators.ddp_cpu_spawn_backend import DDPCPUSpawnBackend
-from pytorch_lightning.accelerators.dp_backend import DataParallelBackend
-from pytorch_lightning.accelerators.gpu_backend import GPUBackend
-from pytorch_lightning.accelerators.tpu_backend import TPUBackend
-from pytorch_lightning.accelerators.horovod_backend import HorovodBackend
-from pytorch_lightning.accelerators.ddp_slurm_backend import DDPSLURMBackend
-from pytorch_lightning.accelerators.ddp_torchelastic_backend import DDPTorchElasticBackend
-from pytorch_lightning.accelerators.ddp_cpu_torchelastic_backend import DDPCPUTorchElasticBackend
-from pytorch_lightning.accelerators.ddp_cpu_slurm_backend import DDPCPUSLURMBackend
-from pytorch_lightning.accelerators.base_accelerator import Accelerator
+from pytorch_lightning.accelerators.cpu_accelerator import CPUAccelerator
+from pytorch_lightning.accelerators.ddp2_accelerator import DDP2Accelerator
+from pytorch_lightning.accelerators.ddp_accelerator import DDPAccelerator
+from pytorch_lightning.accelerators.ddp_spawn_accelerator import DDPSpawnAccelerator
+from pytorch_lightning.accelerators.ddp_cpu_spawn_accelerator import DDPCPUSpawnAccelerator
+from pytorch_lightning.accelerators.dp_accelerator import DataParallelAccelerator
+from pytorch_lightning.accelerators.gpu_accelerator import GPUAccelerator
+from pytorch_lightning.accelerators.tpu_accelerator import TPUAccelerator
+from pytorch_lightning.accelerators.horovod_accelerator import HorovodAccelerator
+from pytorch_lightning.accelerators.ddp_slurm_accelerator import DDPSLURMAccelerator
+from pytorch_lightning.accelerators.ddp_torchelastic_accelerator import DDPTorchElasticAccelerator
+from pytorch_lightning.accelerators.ddp_cpu_torchelastic_accelerator import DDPCPUTorchElasticAccelerator
+from pytorch_lightning.accelerators.ddp_cpu_slurm_accelerator import DDPCPUSLURMAccelerator
+from pytorch_lightning.accelerators.accelerator import Accelerator
