+++ pytorch_lightning/accelerators/ddp_slurm_accelerator.py
@@ -16,7 +16,7 @@ import torch
 import torch.distributed as torch_distrib
 import torch.distributed as dist
 
-from pytorch_lightning.accelerators.base_accelerator import Accelerator
+from pytorch_lightning.accelerators.accelerator import Accelerator
 from pytorch_lightning import _logger as log
 from pytorch_lightning.utilities import AMPType
 from pytorch_lightning.utilities.distributed import rank_zero_only
@@ -25,7 +25,6 @@ from pytorch_lightning.distributed.dist import LightningDistributed
 from pytorch_lightning.overrides.data_parallel import LightningDistributedDataParallel
 from torch.nn.parallel import DistributedDataParallel
 from typing import List
-
 
 try:
     from hydra.utils import to_absolute_path, get_original_cwd
@@ -38,17 +37,17 @@ else:
 
 # -------------------------------------------
 # !!!!!!!!!!!!!! NOTE !!!!!!!!!!!!!!!!!!!!!!
-# TEMP CLASS WHILE WE DECOUPLE TE FROM DDP
+# TEMP CLASS WHILE WE DECOUPLE SLURM FROM DDP
 # !!!!!!!!!!!!!! NOTE !!!!!!!!!!!!!!!!!!!!!!
 # -------------------------------------------
-class DDPCPUSLURMBackend(Accelerator):
+class DDPSLURMAccelerator(Accelerator):
 
     def __init__(self, trainer, cluster_environment=None):
         super().__init__(trainer, cluster_environment)
         self.task_idx = None
         self._has_spawned_children = False
         self.dist = LightningDistributed()
-        self.nickname = 'ddp_cpu'
+        self.nickname = 'ddp'
 
     def setup(self, model):
         self.trainer.model = model
@@ -64,10 +63,12 @@ class DDPCPUSLURMBackend(Accelerator):
         self.trainer.world_size = self.trainer.num_nodes * self.trainer.num_processes
 
     def model_to_device(self, model, process_idx):
-        model.cpu()
+        self.trainer.root_gpu = process_idx
+        torch.cuda.set_device(self.trainer.root_gpu)
+        model.cuda(self.trainer.root_gpu)
 
     def get_device_ids(self):
-        device_ids = None
+        device_ids = [self.trainer.root_gpu]
         return device_ids
 
     def training_step(self, args):
@@ -112,6 +113,10 @@ class DDPCPUSLURMBackend(Accelerator):
         Returns:
 
         """
+        seed = os.environ.get("PL_GLOBAL_SEED")
+        if seed is not None:
+            seed_everything(int(seed))
+
         # determine which process we are and world size
         self.set_world_ranks(process_idx)
 
@@ -138,7 +143,7 @@ class DDPCPUSLURMBackend(Accelerator):
         # on world_size=0 let everyone know training is starting
         if self.trainer.is_global_zero and not torch.distributed.is_initialized():
             log.info('-' * 100)
-            log.info(f'distributed_backend={self.trainer.distributed_backend} (TORCH_ELASTIC)')
+            log.info(f'distributed_backend={self.trainer.distributed_backend} (on SLURM)')
             log.info(f'All DDP processes registered. Starting ddp with {self.trainer.world_size} processes')
             log.info('-' * 100)
