+++ pytorch_lightning/accelerators/ddp_torchelastic_accelerator.py
@@ -16,7 +16,7 @@ import torch
 import torch.distributed as torch_distrib
 import torch.distributed as dist
 
-from pytorch_lightning.accelerators.base_accelerator import Accelerator
+from pytorch_lightning.accelerators.accelerator import Accelerator
 from pytorch_lightning import _logger as log
 from pytorch_lightning.utilities import AMPType
 from pytorch_lightning.utilities.distributed import rank_zero_only
@@ -25,6 +25,7 @@ from pytorch_lightning.distributed.dist import LightningDistributed
 from pytorch_lightning.overrides.data_parallel import LightningDistributedDataParallel
 from torch.nn.parallel import DistributedDataParallel
 from typing import List
+
 
 try:
     from hydra.utils import to_absolute_path, get_original_cwd
@@ -40,7 +41,7 @@ else:
 # TEMP CLASS WHILE WE DECOUPLE SLURM FROM DDP
 # !!!!!!!!!!!!!! NOTE !!!!!!!!!!!!!!!!!!!!!!
 # -------------------------------------------
-class DDPSLURMBackend(Accelerator):
+class DDPTorchElasticAccelerator(Accelerator):
 
     def __init__(self, trainer, cluster_environment=None):
         super().__init__(trainer, cluster_environment)
@@ -51,7 +52,7 @@ class DDPSLURMBackend(Accelerator):
 
     def setup(self, model):
         self.trainer.model = model
-        self.task_idx = int(os.environ['SLURM_LOCALID'])
+        self.task_idx = int(os.environ['LOCAL_RANK'])
 
     def train(self):
         model = self.trainer.model
@@ -113,10 +114,6 @@ class DDPSLURMBackend(Accelerator):
         Returns:
 
         """
-        seed = os.environ.get("PL_GLOBAL_SEED")
-        if seed is not None:
-            seed_everything(int(seed))
-
         # determine which process we are and world size
         self.set_world_ranks(process_idx)
