+++ pytorch_lightning/accelerators/gpu_accelerator.py
@@ -14,12 +14,12 @@
 
 import torch
 
-from pytorch_lightning.accelerators.base_accelerator import Accelerator
+from pytorch_lightning.accelerators.accelerator import Accelerator
 from pytorch_lightning.utilities import AMPType
 from pytorch_lightning.distributed.dist import LightningDistributed
 
 
-class GPUBackend(Accelerator):
+class GPUAccelerator(Accelerator):
     amp_backend: AMPType
 
     def __init__(self, trainer, cluster_environment=None):
